Implemented the menu system! Mostly.
aiosqlite would need an adapter to work with Trio, and there is such an adapter (https://trio-asyncio.readthedocs.io//) but I don't have time to figure that out right now.
So we're back to plain old blocking SQLite. But since we hold an exclusive lock on it, things should be fine.

Didn't get around to implement actually SAVING to the db yet either...

There's also a huge slowdown constructing the loops because constructing the fontmap is slow. However, we should be able to do that just once and reuse it for all the renderers.